+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Highlights are marked with a pancake 🥞
 - Enable nonblocking for unbound socket used in mdns discovery [#794](https://github.com/p2panda/p2panda/pull/794)
 - Allow managing multiple long-term pre-keys [#830](https://github.com/p2panda/p2panda/pull/830)
 - Do not overwrite serde errors during deserialization of `Header` [#886](https://github.com/p2panda/p2panda/pull/886)
+- Handle outdated operations which got processed while being pruned, fix overflow substraction bug [#894](https://github.com/p2panda/p2panda/pull/894)
 
 ## [0.4.0] - 07/07/2025